add update and delete API

+++ src/main/java/petrolimex/model/DMN.java
@@ -18,7 +18,7 @@ public class DMN extends PanacheEntityBase {
     @Id
     @Column(name = "DMN_ID")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    public Integer id;
+    public Long id;
 
     @Column(name = "DMN_NAME")
     public String dmnName;
